requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
lark-oapi==1.2.0
